refactor conceptresultwrapper.hpp
