+++ cpp/lib/answer/conceptmap.cpp
@@ -24,7 +24,7 @@
 #include "typedb/common/exception.hpp"
 
 #include "inc/conceptfactory.hpp"
-#include "inc/conceptresultwrapper.hpp"
+#include "inc/conceptiterator.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
