+++ cpp/lib/concept/thing/attribute.cpp
@@ -28,7 +28,6 @@
 #include "typedb/connection/transaction.hpp"
 
 #include "inc/conceptfactory.hpp"
-#include "inc/conceptresultwrapper.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
