+++ cpp/lib/concept/type/attributetype.cpp
@@ -26,7 +26,6 @@
 #include "typedb/connection/transaction.hpp"
 
 #include "inc/conceptfactory.hpp"
-#include "inc/conceptresultwrapper.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
