+++ cpp/lib/concept/type/relationtype.cpp
@@ -27,7 +27,6 @@
 #include "typedb/connection/transaction.hpp"
 
 #include "inc/conceptfactory.hpp"
-#include "inc/conceptresultwrapper.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
