+++ cpp/lib/inc/conceptfuture.hpp
@@ -1,0 +1,52 @@
+/*
+ * Copyright (C) 2022 Vaticle
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+#pragma once
+#include <vector>
+
+#include "typedb/common/native.hpp"
+#include "typedb/concept/concept.hpp"
+
+namespace TypeDB {
+
+class ConceptFutureWrapper {
+   public:
+    virtual _native::Concept* resolve() = 0;
+};
+
+class ConceptFutureWrapperSimple : public ConceptFutureWrapper {
+   public:
+    virtual _native::Concept* resolve() override;
+    ConceptFutureWrapperSimple(_native::ConceptPromise* promise);
+
+   private:
+    _native::ConceptPromise* nativePromise;
+};
+
+class ConceptFutureWrapperExplicit : public ConceptFutureWrapper {
+   public:
+    virtual _native::Concept* resolve() override;
+    ConceptFutureWrapperExplicit(_native::Concept* concept);
+
+   private:
+    _native::Concept* nativeConcept;
+};
+
+}  // namespace TypeDB
